requirements.txt
aiormq==6.7.7
azure-servicebus==7.11.1
azure-storage-queue==12.6.0